# Conflicts:
#	README.md
#	jshERP-boot/docs/jsh_erp.sql
#	jshERP-boot/pom.xml
#	jshERP-boot/src/main/java/com/jsh/erp/datasource/vo/DepotHeadVo4List.java
#	jshERP-boot/src/main/java/com/jsh/erp/service/DepotHeadService.java
#	jshERP-boot/src/main/java/com/jsh/erp/service/UserService.java
#	jshERP-web/src/components/tools/UserMenu.vue
#	jshERP-web/src/views/bill/AllocationOutList.vue
#	jshERP-web/src/views/bill/OtherInList.vue
#	jshERP-web/src/views/bill/OtherOutList.vue
#	jshERP-web/src/views/bill/PurchaseApplyList.vue
#	jshERP-web/src/views/bill/PurchaseBackList.vue
#	jshERP-web/src/views/bill/PurchaseInList.vue
#	jshERP-web/src/views/bill/PurchaseOrderList.vue
#	jshERP-web/src/views/bill/mixins/BillModalMixin.js
#	jshERP-web/src/views/bill/modules/OtherInModal.vue
#	jshERP-web/src/views/bill/modules/OtherOutModal.vue
#	jshERP-web/src/views/bill/modules/PurchaseInModal.vue
#	jshERP-web/src/views/system/AccountList.vue

+++ jshERP-boot/src/main/java/com/jsh/erp/service/UserService.java
@@ -288,24 +288,27 @@ public class UserService {
      * @param uuid 唯一标识
      * @return 结果
      */
-    public void validateCaptcha(String code, String uuid) {
-        if(StringUtil.isNotEmpty(code) && StringUtil.isNotEmpty(uuid)) {
-            code = code.trim();
-            uuid = uuid.trim();
-            String verifyKey = BusinessConstants.CAPTCHA_CODE_KEY + uuid;
-            String captcha = redisService.getCacheObject(verifyKey);
-            redisService.deleteObject(verifyKey);
-            if (captcha == null) {
-                logger.error("异常码[{}],异常提示[{}]", ExceptionConstants.USER_JCAPTCHA_EXPIRE_CODE, ExceptionConstants.USER_JCAPTCHA_EXPIRE_MSG);
-                throw new BusinessRunTimeException(ExceptionConstants.USER_JCAPTCHA_EXPIRE_CODE, ExceptionConstants.USER_JCAPTCHA_EXPIRE_MSG);
+    public void validateCaptcha(String code, String uuid) throws Exception {
+        PlatformConfig platformConfig = platformConfigService.getInfoByKey("checkcode_flag");
+        if(platformConfig!=null && "1".equals(platformConfig.getPlatformValue())) {
+            if(StringUtil.isNotEmpty(code) && StringUtil.isNotEmpty(uuid)) {
+                code = code.trim();
+                uuid = uuid.trim();
+                String verifyKey = BusinessConstants.CAPTCHA_CODE_KEY + uuid;
+                String captcha = redisService.getCacheObject(verifyKey);
+                redisService.deleteObject(verifyKey);
+                if (captcha == null) {
+                    logger.error("异常码[{}],异常提示[{}]", ExceptionConstants.USER_JCAPTCHA_EXPIRE_CODE, ExceptionConstants.USER_JCAPTCHA_EXPIRE_MSG);
+                    throw new BusinessRunTimeException(ExceptionConstants.USER_JCAPTCHA_EXPIRE_CODE, ExceptionConstants.USER_JCAPTCHA_EXPIRE_MSG);
+                }
+                if (!code.equalsIgnoreCase(captcha)) {
+                    logger.error("异常码[{}],异常提示[{}]", ExceptionConstants.USER_JCAPTCHA_ERROR_CODE, ExceptionConstants.USER_JCAPTCHA_ERROR_MSG);
+                    throw new BusinessRunTimeException(ExceptionConstants.USER_JCAPTCHA_ERROR_CODE, ExceptionConstants.USER_JCAPTCHA_ERROR_MSG);
+                }
+            } else {
+                logger.error("异常码[{}],异常提示[{}]", ExceptionConstants.USER_JCAPTCHA_EMPTY_CODE, ExceptionConstants.USER_JCAPTCHA_EMPTY_MSG);
+                throw new BusinessRunTimeException(ExceptionConstants.USER_JCAPTCHA_EMPTY_CODE, ExceptionConstants.USER_JCAPTCHA_EMPTY_MSG);
             }
-            if (!code.equalsIgnoreCase(captcha)) {
-                logger.error("异常码[{}],异常提示[{}]", ExceptionConstants.USER_JCAPTCHA_ERROR_CODE, ExceptionConstants.USER_JCAPTCHA_ERROR_MSG);
-                throw new BusinessRunTimeException(ExceptionConstants.USER_JCAPTCHA_ERROR_CODE, ExceptionConstants.USER_JCAPTCHA_ERROR_MSG);
-            }
-        } else {
-            logger.error("异常码[{}],异常提示[{}]", ExceptionConstants.USER_JCAPTCHA_EMPTY_CODE, ExceptionConstants.USER_JCAPTCHA_EMPTY_MSG);
-            throw new BusinessRunTimeException(ExceptionConstants.USER_JCAPTCHA_EMPTY_CODE, ExceptionConstants.USER_JCAPTCHA_EMPTY_MSG);
         }
     }
 
@@ -378,16 +381,8 @@ public class UserService {
             user.setPassword(null);
             // Todo childs 注释 不发微信订阅消息
 //            if(BusinessConstants.DEFAULT_MANAGER.equals(user.getLoginName())) {
-//                //如果是管理员，则发送订阅消息
-//                //1-获取token
-//                String accessToken = platformConfigService.getAccessToken();
-//                //2-发送订阅消息
-//                String templateId = platformConfigService.getPlatformConfigByKey("login_temp_id").getPlatformValue();
-//                String weixinUrl = platformConfigService.getPlatformConfigByKey("weixinUrl").getPlatformValue();
-//                String platformName = platformConfigService.getPlatformConfigByKey("platform_name").getPlatformValue();
-//                if(StringUtil.isNotEmpty(accessToken) && StringUtil.isNotEmpty(user.getWeixinOpenId()) && StringUtil.isNotEmpty(templateId)) {
-//                    platformConfigService.sendSubscribeMessage(accessToken, weixinUrl, platformName, templateId, null, user.getWeixinOpenId());
-//                }
+//                //如果是管理员，则发送登录邮件
+//                sendEmailToCurrentUser(request, user);
 //            }
             redisService.storageObjectBySession(token,"clientIp", Tools.getLocalIp(request));
             logService.insertLogWithUserId(user.getId(), user.getTenantId(), "用户",
@@ -458,6 +453,26 @@ public class UserService {
             user = list.get(0);
         }
         return user;
+    }
+
+    /**
+     * 发送邮件给当前用户
+     * @param request
+     * @param user
+     * @throws Exception
+     */
+    private void sendEmailToCurrentUser(HttpServletRequest request, User user) throws Exception {
+        String platformName = platformConfigService.getPlatformConfigByKey("platform_name").getPlatformValue();
+        String emailFrom = platformConfigService.getPlatformConfigByKey("email_from").getPlatformValue();
+        String emailAuthCode = platformConfigService.getPlatformConfigByKey("email_auth_code").getPlatformValue();
+        String emailSmtpHost = platformConfigService.getPlatformConfigByKey("email_smtp_host").getPlatformValue();
+        if(StringUtil.isNotEmpty(emailFrom) && StringUtil.isNotEmpty(emailAuthCode) && StringUtil.isNotEmpty(emailSmtpHost)
+                && StringUtil.isNotEmpty(user.getEmail())) {
+            String emailSubject = "用户" + user.getLoginName() + "成功登录" + platformName;
+            String emailBody = "用户" + user.getLoginName() + "成功登录" + platformName + "，登录时间：" + Tools.getCenternTime(new Date())
+                    + "，登录IP：" + Tools.getLocalIp(request);
+            platformConfigService.sendEmail(emailFrom, emailAuthCode, emailSmtpHost, user.getEmail(), emailSubject, emailBody);
+        }
     }
 
     public int checkIsNameExist(Long id, String name)throws Exception {
